finish moving source_modules to otf.source
