+++ src/functional/otf/compilation/build_systems/cmake_lists.py
@@ -3,8 +3,8 @@ from typing import Sequence
 
 import eve
 from eve.codegen import JinjaTemplate as as_jinja
-from functional.otf import source
 from functional.otf.compilation import common
+from functional.otf.source import source
 
 
 class FindDependency(eve.Node):
