--- src/functional/program_processors/source_modules/__init__.py
@@ -1,18 +1,0 @@
-from .source_modules import (
-    BindingModule,
-    BufferParameter,
-    Function,
-    LibraryDependency,
-    ScalarParameter,
-    SourceModule,
-)
-
-
-__all__ = [
-    "SourceModule",
-    "BindingModule",
-    "Function",
-    "ScalarParameter",
-    "BufferParameter",
-    "LibraryDependency",
-]
